+++ src/main/scala/org/mmisw/orr/ont/app/OntController.scala
@@ -12,7 +12,7 @@ import org.json4s.native.JsonMethods._
 import org.mmisw.orr.ont._
 import org.mmisw.orr.ont.db.{Ontology, OntologyVersion}
 import org.mmisw.orr.ont.service._
-import org.mmisw.orr.ont.swld.{V2RModel, ontUtil}
+import org.mmisw.orr.ont.swld.{M2RModel, V2RModel, ontUtil}
 import org.scalatra.Created
 import org.scalatra.servlet.{FileItem, FileUploadSupport, SizeConstraintExceededException}
 
@@ -209,6 +209,12 @@ class OntController(implicit setup: Setup,
         val newV2r = oldV2r.copy(metadata = Some(newMetadata))
         java.nio.file.Files.write(destFile.toPath,
           newV2r.toPrettyJson.getBytes(java.nio.charset.StandardCharsets.UTF_8))
+      }
+      else if (actualFormat == "m2r") {
+        val oldM2r = parse(file).extract[M2RModel]
+        val newM2r = oldM2r.copy(metadata = Some(newMetadata))
+        java.nio.file.Files.write(destFile.toPath,
+          newM2r.toPrettyJson.getBytes(java.nio.charset.StandardCharsets.UTF_8))
       }
       else {
         val ontModel = ontUtil.loadOntModel(uri, file, actualFormat)