+++ ChangeLog.md
@@ -1,5 +1,8 @@
 ## change log ##
 
+* 2016-05-16: 0.3.1:
+  - initial version of m2r format for ORR mapping ontologies
+  
 * 2016-05-12: 0.3.1:
   - Ontology model now with required ownerName entry instead of optional orgName.
     The syntax for this string: `ownerName ::= <orgName> | ~<userName>`.